Added crc check

+++ obc/bl/include/bl_config.h
@@ -11,7 +11,7 @@
 //
 // The flash image of the bootloader must not be larger than this value.
 //*****************************************************************************
-#define APP_START_ADDRESS (uint32_t)0x00040000
+#define APP_START_ADDRESS 0x00040000
 
 // The start address of metadata, and the size (used in bl_flash.c and bl_main.c)
